feat: `unsafe { … }` 有类型了 —— 而且是执行器一直以来给的那个
类型检查器里写着:块自己的类型"暂时保持 Any,和 Expr::Block 一致;一个求值为有类型
的值的块是另一件事"。执行器从来没同意过 —— 它早就把 `unsafe` 块求值成最后一条语句
的值,带分号也一样:`unsafe { 7; }` 就是 7。现在类型追上了那个值。

`Any` 会扩散,而它扩散到的地方正好就是这个 demo 由之构成的那种代码。量一个中断桩
数组的步长,是两个地址相减;两边都是 `Any` 的时候,结果没有任何 `as Int` 出得去,而
报错指的是那个转型而不是缺失的类型。这件事在这几轮里咬了我两次。

**这个修复的另一半**:`symbol_address` 和 `call_address_2` 在类型检查器里根本没有
条目,于是调用它们产出 `Any`,而且既不检查元数也不检查参数。现在有了。而且
`symbol_address` 还坚持它的名字必须是*字面量* —— 一个重定位就是在镜像链接时解析掉的
名字,运行期没有任何东西可以查。传一个变量以前能过类型检查、在 VM 上运行(被拒绝)、
然后以一条关于操作码的消息在原生降低时失败。

**把这个洞堵上,立刻抓到两个真的错误,都在这个仓库自己的测试夹具里:**

    let a = unsafe { volatile_read_u32(reg) };
    let b = unsafe { volatile_read_u32(reg) };
    return a + b;              // 而这个函数返回 Int

一次 32 位读就是一个 `u32`,两个相加在有人说明之前不是 `Int`。驱动们一直都写着那个
转型;夹具没有,因为 `Any` 不逼它们写。

`bare-metal-x86` 里那些"只是为了拿到类型"的 `let x: Int = unsafe { … }` 标注随之
去掉 —— 它们本来就是在给一个不该丢失的类型打补丁。

裸机十二项检查全绿;工作区测试全过;AOT 覆盖率 48/51 不变;性能几何平均 1.006x。

+++ cli/tests/clif_differential_test.rs
@@ -637,8 +637,8 @@ fn volatile_reads_are_not_collapsed() {
         &source,
         "fn read_twice(addr: usize) -> Int {\n\
          \x20   let reg = addr as *mut u32;\n\
-         \x20   let a = unsafe { volatile_read_u32(reg) };\n\
-         \x20   let b = unsafe { volatile_read_u32(reg) };\n\
+         \x20   let a = unsafe { volatile_read_u32(reg) } as Int;\n\
+         \x20   let b = unsafe { volatile_read_u32(reg) } as Int;\n\
          \x20   return a + b;\n\
          }\n\
          return read_twice(0x1000);\n",
@@ -792,7 +792,7 @@ fn critical_section_emits_its_instructions_in_order() {
          \x20   let saved = unsafe { cpu_irq_save() };\n\
          \x20   unsafe { volatile_write_u32(reg, 1 as u32); };\n\
          \x20   unsafe { cpu_barrier(); };\n\
-         \x20   let v = unsafe { volatile_read_u32(reg) };\n\
+         \x20   let v = unsafe { volatile_read_u32(reg) } as Int;\n\
          \x20   unsafe { cpu_irq_restore(saved); };\n\
          \x20   return v;\n\
          }\n\
